+++ internal/repository/seller/seller_interface.go
