+++ internal/router/sellers.go
@@ -2,10 +2,10 @@ package router
 
 import (
 	"github.com/go-chi/chi/v5"
-	"github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler"
+	sellerHandler "github.com/varobledo_meli/W17-G10-Bootcamp.git/internal/handler/seller"
 )
 
-func MountSellerRoutes(api chi.Router, hd *handler.SellerHandler) {
+func MountSellerRoutes(api chi.Router, hd *sellerHandler.SellerHandler) {
 	api.Route("/sellers", func(r chi.Router) {
 		r.Get("/", hd.FindAll)
 		r.Get("/{id}", hd.FindById)
